move text specific logic to subclass
